Reorganize download code to prepare renaming module
