+++ src/lib/index.ts
@@ -2,7 +2,7 @@ import getBody from './getBody';
 import getTracks from './getTracks';
 import validateTracks from './validateTracks';
 import logger from './logger';
-import downloadWithConcurrencyLimit from './downloadWithConcurrencyLimit';
+import downloadWithConcurrencyLimit from './downloadTracks/downloadWithConcurrencyLimit';
 
 export {
   getBody, getTracks, validateTracks, logger, downloadWithConcurrencyLimit,
